power-platform/backend/app/services/hil/ni_crio_stub.py
"""NI cRIO gRPC stub adapter."""
from typing import Dict, List, Any
from .base import BaseHILAdapter, HILChannel, HILTelemetry, HILStatus
import time

# This is a stub - full implementation requires gRPC client code
# and proto definitions specific to your NI cRIO setup

class NI_cRIO_Adapter(BaseHILAdapter):
    """
    National Instruments CompactRIO adapter via gRPC.

    STUB IMPLEMENTATION - Requires:
    1. gRPC proto definitions for your cRIO VI
    2. Generated Python client code
    3. cRIO network configuration
    """

    def __init__(self, config: Dict[str, Any]):
        """
        Initialize NI cRIO adapter.

        Config:
            crio_host: cRIO IP address
            crio_port: gRPC port (typically 50051)
            vi_name: LabVIEW VI name
        """
        super().__init__(config)
        self.crio_host = config.get("crio_host", "192.168.1.100")
        self.crio_port = config.get("crio_port", 50051)
        self.vi_name = config.get("vi_name", "HIL_Interface.vi")

    async def connect(self) -> bool:
        """
        Connect to cRIO via gRPC.

        Implementation requires:
            import grpc
            from generated import crio_pb2, crio_pb2_grpc

            channel = grpc.aio.insecure_channel(f'{self.crio_host}:{self.crio_port}')
            self.stub = crio_pb2_grpc.HILServiceStub(channel)
        """
        self.status = HILStatus.CONNECTED
        return True

    async def disconnect(self) -> bool:
        """Disconnect from cRIO."""
        self.status = HILStatus.DISCONNECTED
        return True

    async def configure_channels(self, channels: List[HILChannel]) -> bool:
        """Configure cRIO I/O channels."""
        for ch in channels:
            self.channels[ch.name] = ch
        # Would call gRPC ConfigureChannels method
        return True

    async def start_stream(self, sample_rate: float = 1000.0) -> bool:
        """Start cRIO data acquisition."""
        self.status = HILStatus.RUNNING
        # Would call gRPC StartAcquisition method
        return True

    async def stop_stream(self) -> bool:
        """Stop cRIO data acquisition."""
        self.status = HILStatus.CONNECTED
        # Would call gRPC StopAcquisition method
        return True

    async def read_channels(self) -> HILTelemetry:
        """Read from cRIO via gRPC streaming."""
        # Would call gRPC ReadTelemetry streaming method
        return HILTelemetry(
            timestamp=time.time(),
            channels={},
            faults=[],
            status=self.status
        )

    async def write_setpoints(self, setpoints: Dict[str, float]) -> bool:
        """Write setpoints to cRIO."""
        # Would call gRPC WriteOutputs method
        return True


"""
Required proto file (crio.proto):

syntax = "proto3";

service HILService {
    rpc ConfigureChannels(ChannelConfig) returns (StatusResponse);
    rpc StartAcquisition(AcqConfig) returns (StatusResponse);
    rpc StopAcquisition(Empty) returns (StatusResponse);
    rpc ReadTelemetry(Empty) returns (stream TelemetryData);
    rpc WriteOutputs(OutputData) returns (StatusResponse);
}

message ChannelConfig {
    repeated Channel channels = 1;
}

message Channel {
    string name = 1;
    string type = 2;
    double range_min = 3;
    double range_max = 4;
}

message AcqConfig {
    double sample_rate = 1;
}

message TelemetryData {
    double timestamp = 1;
    map<string, double> values = 2;
}

message OutputData {
    map<string, double> setpoints = 1;
}

message StatusResponse {
    bool success = 1;
    string message = 2;
}

message Empty {}

Generate with: python -m grpc_tools.protoc -I. --python_out=. --grpc_python_out=. crio.proto
"""
