power-platform/backend/pyproject.toml
[project]
name = "power-platform-backend"
version = "1.0.0"
description = "Cloud-Native Power Electronics Platform - Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.1",
    "celery>=5.3.4",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "authlib>=1.3.0",
    "httpx>=0.26.0",
    "numpy>=1.26.3",
    "scipy>=1.11.4",
    "matplotlib>=3.8.2",
    "plotly>=5.18.0",
    "pandas>=2.1.4",
    "pyyaml>=6.0.1",
    "weasyprint>=60.2",
    "jinja2>=3.1.3",
    "python-socketio>=5.11.0",
    "asyncio-mqtt>=0.16.2",
    "pymodbus>=3.6.4",
    "opcua>=0.98.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
